Reorganização de classes
Movida as classes de modelagem de Associado e Convenio para a pasta Model
