+++ src/Control/CCASS.java
@@ -1,5 +1,6 @@
 package Control;
 
+import Model.Associado;
 import javafx.application.Application;
 import javafx.event.EventHandler;
 import javafx.fxml.FXMLLoader;
