+++ src/Control/CadastraCliente.java
@@ -1,5 +1,6 @@
 package Control;
 
+import Model.Associado;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
@@ -9,7 +10,6 @@ import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Optional;
 
 /**
