+++ src/Control/CadastraConvenio.java
@@ -1,5 +1,6 @@
 package Control;
 
+import Model.Convenio;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
