+++ src/Control/ControladoraPrincipal.java
@@ -1,5 +1,6 @@
 package Control;
 
+import Model.Associado;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
