+++ src/Model/Associado.java
@@ -1,8 +1,7 @@
-package Control;
+package Model;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
-import javafx.scene.control.Button;
 
 import java.util.ArrayList;
 
@@ -22,7 +21,6 @@ public class Associado {
     private String CPF;
     private String Tel;
     private String Endereco;
-    private Button Imprimir = new Button("dummy");
     private ArrayList<Dependente> Dependentes = new ArrayList<Dependente>();
 
     public Associado(String nome, String CPF, String RG, String Tel, String endereco, String Pai, String Mae, ArrayList<Dependente> dependentes) {
